moved overload check to BugsnagUnity.Tests
