+++ features/fixtures/maze_runner/Assets/Scripts/Main.cs
@@ -139,35 +139,6 @@ public class Main : MonoBehaviour
 #endif
     }
 
-
-
-    // this method never runs and is just used to check that no notify overides are accidentally broken during refactoring
-    // if one is broken then the fixture will not compile
-    private void OverloadCheck()
-    {
-
-        Bugsnag.Notify("name", "message", "stacktrace");
-
-        Bugsnag.Notify("name", "message", "stacktrace", CallBack);
-
-        Bugsnag.Notify(new Exception());
-
-        Bugsnag.Notify(new Exception(), "stacktrace");
-
-        Bugsnag.Notify(new Exception(), "stacktrace", CallBack);
-
-        Bugsnag.Notify(new Exception(), CallBack);
-
-        Bugsnag.Notify(new Exception(), Severity.Error);
-
-        Bugsnag.Notify(new Exception(), Severity.Error, CallBack);
-    }
-
-    private bool CallBack(IEvent e)
-    {
-        return true;
-    }
-
 }
